Added number checking, score adding and text removing
I’ve added number checking, which will check the inputed number to see
if it’s the right answer, score adding, which will add point to the
score if it’s right =, and text removing, which will remove the text in
the entry box (text box) when the button is clicked and the answer is
right.
Note: The answer to all the questions is given in the console.